@@ -4,7 +4,7 @@ version '1.0-SNAPSHOT'
 apply plugin: 'java'
 apply plugin: 'antlr'
 
-sourceCompatibility = 1.8
+sourceCompatibility = 10
 
 repositories {
     mavenCentral()
@@ -21,6 +21,11 @@ dependencies {
     compile 'io.javaslang:javaslang-match:2.0.5'
 
 
-    testCompile "org.slf4j:slf4j-simple:1.7.25"
-    testCompile group: 'junit', name: 'junit', version: '4.12'
+    testImplementation(
+            'org.junit.jupiter:junit-jupiter-api:5.1.0'
+    )
+    testRuntimeOnly(
+            'org.slf4j:slf4j-simple:1.7.25',
+            'org.junit.jupiter:junit-jupiter-engine:5.1.0'
+    )
 }
